+++ examples/demo_smollm_chat.py
@@ -4,7 +4,27 @@ from demo_models import models
 from demo_messages import selfaware_consciousness_messages as messages
 
 
-def demo(model: Model):
+def demo_prompt(model: Model):
+    print(model)
+    config = get_config(model.creator_hf_repo)
+
+    options = Options(
+        ctx_size=config.max_position_embeddings,
+        predict=-2,
+        model=model,
+        prompt='Explain meaning of life.',
+        no_display_prompt=True,
+        gpu_layers=99,
+    )
+
+    for chunk in llama_generate(options):
+        print(chunk, flush=True, end='')
+
+    # newline
+    print('\n' + '-' * 20)
+
+
+def demo_messages(model: Model):
     print(model)
     config = get_config(model.creator_hf_repo)
 
@@ -26,9 +46,11 @@ def demo(model: Model):
 
 if __name__ == '__main__':
     models_ids: list[str] = [
+        'HuggingFaceTB/SmolLM2-360M-Instruct',
         'HuggingFaceTB/SmolLM2-1.7B-Instruct',
     ]
 
     for model_id in models_ids:
         model: Model = models[model_id]
-        demo(model)
+        demo_prompt(model)
+        demo_messages(model)